Rename Waterskipper ralated files, and update includes
Ant.h/cpp renamed to Waterskipper.h/cpp
AntHill/h/cpp renamed to WaterskipperNest.h/cpp
